### 2020-05-01 #38 @MelanieIStefan
Ten-year challenge: Reproduction of An allosteric model of calmodulin [PNAS 105(31), 10768 (2008)]
**Original article:** M.I. Stefan, S.J. Edelstein, N. Le Novère (2008). An allosteric model of calmodulin explains differential activation of PP2B and CaMKII. PNAS 105(31), 10768.

**Metadata URL:**

**Scientific domain:**  Systems neuroscience
**Programming language:**  SBML, R
**Suggested editor:**  This is part of the 10-year challenge, so should be handled by the challenge organisers
