### 2020-05-04 #38 @rougier
Thanks for your submission, I will edit it.

@apdavison Could you review this submission for the Ten Years Reproducibility Challenge?
