### 2020-06-17 #38 @apdavison
Hi @MelanieIStefan I'm struggling a bit with this, as I haven't used Copasi before. I've got as far as loading "Figure3.cps" and running a parameter scan, but I can't figure out how to save the results (i.e. to try to recreate Figure3.txt). Please can you help me out with a detailed description of what to click on/select?
