### 2020-06-19 #38 @MelanieIStefan
Hi @apdavison - when you run the parameter scan, a separate window should be created by Copasi where the results are plotted.  In that window, you can select "Save Data" from the menu above the plot.

![Save](https://user-images.githubusercontent.com/7545731/85159497-80239000-b255-11ea-9dea-7b70b3ef0c0a.png)

This should bring up a menu allowing you to save the data as a .txt file. Please let me know if this doesn't work.
