### 2020-07-06 #38 @apdavison
This is a well-written article, which serves as an excellent example of why reproducibility of modelling studies requires not only the model specification but also the simulation configuration and data analysis scripts to be stored.

I was able to re-run the simulations, and reproduce the data files in the repository.

### Major issues:
- the curves in the manuscript differ substantially from those in the corrections to the original manuscript. The following two graphs compare the curves digitised from the corrected Figure 3 and 4 in Stefan et al. 2008 with the data from the reproduction.

![comparison_figure3](https://user-images.githubusercontent.com/260552/86591000-fa138300-bf90-11ea-89b7-ce9ff07cb60f.png)

![comparison_figure4](https://user-images.githubusercontent.com/260552/86591014-026bbe00-bf91-11ea-9142-89c4379fbcde.png)

On the basis of this, I don't think that the original parameters have been correctly recovered. I recommend an attempt is made to obtain a closer fit to the original curves (perhaps using automated parameter fitting). This might also result in a successful reproduction of Figure 5. Otherwise, I think this has to be labelled an unsuccessful reproduction and the manuscript amended accordingly.

### Minor issues:
- IUPAC recommendations are that the names of elements should not be capitalised, so "Calcium" should be replaced by "calcium" (as in the original article that is being reproduced).
- "an attempt to reproduce the code" should be something like "an attempt to  reproduce the results using the original code"
- "Calcium frequencies" - it is not clear what is meant here. Concentrations?
